backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
firebase-admin==6.5.0
python-dotenv==1.0.1
httpx==0.27.2